Upgrade to UE5
编译一次要30分钟，希望这是最后一次

+++ Source/UEImgui/Private/ImguiWrap/ImguiUEWrap.cpp
@@ -123,7 +123,7 @@ void ImGui::BeginDetail()
 	if (CurrentDetailWidget.IsValid())
 	{
 		FVector2D CurrentDetailSize = CurrentDetailWidget.Pin()->GetCachedGeometry().Size;
-		ImGui::SetNextWindowSizeConstraints({CurrentDetailSize.X, -1}, {CurrentDetailSize.X, -1});
+		ImGui::SetNextWindowSizeConstraints({(float)CurrentDetailSize.X, -1}, {(float)CurrentDetailSize.X, -1});
 	}
 	ImGui::Begin(TCHAR_TO_UTF8(*CurrentDetail.ToString()), nullptr, ImGuiWindowFlags_UEDetail);
 }
